Improved overall organization of source.
